# blogger-apis/api/swagger/swagger1.yaml
swagger: '2.0'
info:
    version: 0.0.1
    title: 'Mean Stack Blogger'
host: 'localhost:10010'
basePath: /
schemes:
    - http
    - https
consumes:
    - application/json
produces:
    - application/json
paths:
...
    '/member/{id}':
        x-swagger-router-controller: member
        get:
            operationId: getOne
            description: 'get a member'
            parameters: [{name: id, type: string, in: path, required: true}]
            responses: {'200': {description: Success, schema: {$ref: '#/definitions/GetMemberResponse'}}, default: {description: Error, schema: {$ref: '#/definitions/ErrorResponse'}}}
        put:
            operationId: update
            description: 'update a member'
            parameters: [{name: id, description: 'Member id', type: string, in: path, required: true}, {name: firstName, description: 'Member properties', in: body, required: true, schema: {$ref: '#/definitions/Member'}}]
            responses: {'200': {description: Success, schema: {$ref: '#/definitions/GeneralResponse'}}, default: {description: Error, schema: {$ref: '#/definitions/ErrorResponse'}}}
        delete:
            operationId: delMember
            description: 'delete a member'
            parameters: [{name: id, description: 'Member id', type: string, in: path, required: true}]
            responses: {'200': {description: Success, schema: {$ref: '#/definitions/GeneralResponse'}}, default: {description: Error, schema: {$ref: '#/definitions/ErrorResponse'}}}
definitions:
...
    ErrorResponse:
        required:
            - message
        properties:
            message: {type: string}
...
    GeneralResponse:
        type: object
        properties:
            success: {type: number, description: 'returns 1 if successful'}
            description: {type: string, description: 'a short comment'}
        required:
            - success
            - description
...
    GetMemberResponse:   
        required:
            - id
            - firstName
            - lastName
            - middleName
            - nationalId
            - email
            - phoneNumber
        properties:
            id       :   {type: number}
            firstName:   {type: string}
            lastName:    {type: string}
            middleName:  {type: string} 
            nationalId:  {type: string}
            email:       {type: string}
            phoneNumber: {type: string}    
...
    Member:
        type: object
        properties:
            firstName: {type: string, description: 'Member First Name'}
            lastName: {type: string, description: 'Member Last Name'}
            middleName: {type: string, description: 'Member Middle Name'}
            dateOfBirth: {type: string, description: 'Member Date of Birth'}
            nationalId: {type: string, description: 'National ID'}
            phoneNumber: {type: string, description: 'Member Phone Number'}
            email: {type: string, description: 'Member Email'}
        required:
            - firstName
            - lastName
            - middleName
            - nationalId
            - email
            - phoneNumber